Added gen_repre features option + ycbv configs
Added configs for YCBV dataset - for template generation and generating of object representation.
Modified `gen_repre.py` to support not only the masked BOW approach from the original paper,
but also the direct feature extraction on the templates. This `features` approach does not take the object
mask into account, but extracts features from the whole image and does not perform registration, pca, knn and tfidf.